task1/ecommerce_project/requirements.txt
Django==3.2
djangorestframework==3.12.4
requests==2.26.0